requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
sqlite3-utils==3.35
pandas==2.1.4
selenium==4.15.2
webdriver-manager==4.0.1
python-dotenv==1.0.0
tqdm==4.66.1
schedule==1.2.0
